Move Playwright install to fixture
Install Playwright in a fixture shared by the UI tests.
