+++ tests/AdventOfCode.Tests/Api/PlaywrightFixture.cs
@@ -1,0 +1,25 @@
+﻿// Copyright (c) Martin Costello, 2015. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+
+namespace MartinCostello.AdventOfCode.Api;
+
+public class PlaywrightFixture : IAsyncLifetime
+{
+    public Task DisposeAsync() => Task.CompletedTask;
+
+    public Task InitializeAsync()
+    {
+        InstallPlaywright();
+        return Task.CompletedTask;
+    }
+
+    private static void InstallPlaywright()
+    {
+        int exitCode = Microsoft.Playwright.Program.Main(new[] { "install" });
+
+        if (exitCode != 0)
+        {
+            throw new InvalidOperationException($"Playwright exited with code {exitCode}.");
+        }
+    }
+}
